[core,shell] introduce ShellIO to all tasks and required protocol
ShellIO introduces indirection which allows for remote shells to open, and
operate on local files.

This lets tasks open and write files from where the shell is running, and not
just on the server.

I/O is heavily buffered (eagerly on 64k), to reduce the total overhead that is
caused by the network.