[all] Convert BackendKeyClause into a more structured BackendKeyFilter
BackendKeyFilter differs from Clause in that it is an immutable carrier with
specific slots for start, end, and limit.

This prevents ambigiouty with multiple start or end clauses (which is
prohibited by CQL), now only one clause can be in effect at a given time.

BackendKeyFilter is immutable, any modifications happens using the #with*
mutators which returns a modified copy.