[all] remove cadence hint in Aggregation
This simplifies the Aggregation interface by removing the initial cadence hint.

This hint is not necessary since the final cadence can be extracted from the
`AggregationInstance` which takes the initial range-based cadence into account
as a hint.

This does slightly change the behaviour of queries like: `sum(10m) | average`,
after this change average will use the resolution derived from the range,
rather than the previous aggregation. I believe this behaviour is more
expected.