Query log - configurable logging of queries
This adds the possibility to log queries, at 1) when they arrive to Heroic, and 2) when they're done. The original query is logged, in addition to information such as a query UUID, ip:port, client ID, elapsed time, number of resulting data points.
Define two new Loggers "query.acccess.log" and "query.done.log" to get the output to separate files. Log level for the query logging is "TRACE".
Query logging is disabled by default. Use Feature flag LOG_QUERIES to enable.
Logging of completed query can be filtered on the amount of data points being returned by the query. Use logQueriesThresholdDataPoints in the metrics section of config file to define lower threshold. Default threshold is 1000000. Sample configuration:

metrics:
  logQueriesThresholdDataPoints: 1000000

We already had a query.log, it was renamed ti query.done.log
Now there's also a query.access.log where incoming queries are printed, together with a query-unique UUID so that query.access.log and query.done.log can be matched with each other. If a node dies, it should now be possible to know which queries that were active/outstanding when it happened.

log4j details:
* Change: query.log renamed to query.done.log
* New Logger: query.access.log