Log incoming query before processing it
We already had a query.log, it was renamed ti query.done.log
Now there's also a query.access.log where incoming queries are printed, together with a query-unique UUID so that query.access.log and query.done.log can be matched with each other. If a node dies, it should now be possible to know which queries that were active/outstanding when it happened.

log4j details:
* Change: query.log renamed to query.done.log
* New Logger: query.access.log