[core] Faster shutdown of Heroic
This change affects the shutdown behaviour of the core ScheduledExecutorService.
It now doesn't wait for scheduled/delayed tasks that aren't yet running but are
scheduled to run at some time in the future.
Previous to this the cluster refresh in clustered configurations would always be
waited upon, introducing a delay in the shutdown.