[all] Read data from MetricBackend continously
When reading data from MetricBackend, read with a consumer in a streamed
fashion. This limits the amount of memory retained during the operation.

This feature is disabled by default. Enable by specifying the
'com.spotify.heroic.sliced_data_fetch' feature flag, either per-query or
in the configuration file, like so:

features:
  - com.spotify.heroic.sliced_data_fetch