[component] fix sampling in bucket aggregations
This modifies bucket aggregations so that they don't include the first bucket in the range since it
will never be provided data due to how it is queried. Data fetches are performed in the range
`(start - extent, start]`, but buckets are setup in the range `[start, end]` with size being used as
a distance between buckets.

Since the bucket selection algorithm needed to change because of this, it was simplified and test
cases added to clarify how it works.