[all] A shard with no discovered nodes is an error
This requires configuration - reusing the already existing 'topology'
configuration variable. Use it to specify which shards that are
expected to exist / be represented by nodes in the cluster. If Heroic
at some point has zero known nodes for any of these shards that will
be treated as a shard error and shown in query responses in the form
of entries in the 'errors' hash.
The list of expected shards is optional, but can be specified as so:

cluster:
  tags:
    site: site2
    part: a
  discovery:
    type: ...
  topology:
    - site: site1
    - site: site2
      part: a
    - site: site2
      part: b
    - site: site3