[component, core] Protect shard rpc operations
This commit protects all rpc calls so that the setup of the call is done
under a lock.
There were previously a race condition where ClusterShard would get a
list of nodes from the ClusterManager and then iterate over the list,
calling the nodes until one of them succeeded. Nothing guaranteed that
the node objects were still valid when the rpc operation was tried. The
node object could have been deemed failed and shut down. This would
result in a cancelled future - unexpected behaviour.