[elasticsearch] Slow-start mode for rate limiting
Existing functionality: The option writesPerSecond enables a rate limit
of how many writes that may be performed per second. Any additional
write requests will be dropped.
New: This commit adds a possibility to have a slow-start mode for this
rate limit, causing the limit to be 3 times lower for the x first
seconds.
Disabled by deafult. Controlled by the new rateLimitSlowStartSeconds
option for metadata and suggest.