[all] Add metrics: read rate & metric density
query-metrics-read-rate - Number of samples read per second from metric
backend, on data-node.
This makes for a latency measurement that is comparable across queries.
This can be used to show if any query loaded the system more than what
would have been expected from just the number of samples.

query-read-rate - Same as above, but on a global query level. Therefore
this also includes network latencies etc, so becomes a measure of the
performance of a larger part of the system.

query-metrics-row-density - Distance between metrics, in milliseconds,
in the data fetched from metric backend for a query.
Since this is on a row level, this gives insights into the density of
the underlying data, no matter how many hosts the query matched etc.
- This metrics was also added in the Query Log output as dataDensity
in outgoing-response-at-node.