[component,core] Add darkload functionality
When testing a new configuration option, backend type, code change, etc
it can be very good to be able to test with the same query load as your
main production cluster. This commit adds this possibility.
Set magic tag darkload:<anything> on your shard. This will cause queries to
fan out to it, but the result will be discarded.