[metadata,suggest] Fix slow start
Instead of using the Guava SmoothWarmingUp implementation, use the
simple SmoothBursty and apply a manual slow start period to it in a
wrapper. Slow start duration is still configurable with
rateLimitSlowStartSeconds. The amount of slow start is hard coded to
factor 4. During startup the rate will be 4 times lower for x seconds.