[heroic-elasticsearch-utils] add DistributedRateLimitedCache
A distributed cache that does not allow to be called more than a specific rate.

Instead of having a cache that is local to each consumer this allows the cache to be distributed to all the instances. This became necessary when migrating from Kafka -> Google PubSub as a consumer with PubSub will not have a stable partition of metrics.

Reads from the distributed cache are stored locally to reduce the load on the distributed cache cluster.