Upgrade elasticsearch client to 7.5 (#594)
Upgrade the ES client to 7.5 which is the latest version. This is the last major version that will support the transport client.

The major change from 5.x -> 7.x is indexes no longer support multiple types.

suggest has two types, series and tags. This means there will be 2x the number of indexes.
metadata has only one type, metadata. There will be the same number of indexes.
Other changes made as part of the PR.

Remove deprecated configuration options.
Completely remove the StandaloneClient as it was removed in 6.x+. TestContainers are now used for the integration tests.
Completely remove the NodeClient since it's not recommended to use anyway.

WARNING this is a breaking change and should be versioned as such. The 7.5x will not work with a 5.x cluster.