[metadata] clear search scroll when complete (#608)
The search scroll must be closed when it's no longer needed otherwise elasticsearch will hold onto them until the timeout is reached. In previous versions of elasticsearch the number of scrolls that could be open was unlimited, with ES7 the default is now 500. It can be overriden with `search.max_open_scroll_context`.

https://www.elastic.co/guide/en/elasticsearch/reference/6.8/search-request-scroll.html#_clear_scroll_api