[heroic-component] remove warnings around deprecated methods
The annotation is enough to alert the developer (us) to move away from the deprecated methods.

I tried to spend some time migrating the QueryTrace methods but had a hard time following all the involved code. Since we want to do a heroic release soon this will at least cut down on the amount of logging being done in production internally.